requirements.txt
djangorestframework-simplejwt
pillow
ruff